# Vectorize `find_first_of` for 8 and 16 bit elements with SSE4.2 `pcmpestri`