--- a/benchmarks/src/find_first_of.cpp
+++ b/benchmarks/src/find_first_of.cpp
@@ -17,7 +17,7 @@ void bm(benchmark::State& state) {
 
     static_assert(Pos < HSize);
     static_assert(Which < NSize);
-    h.at(Pos) = n.at(Which);
+    h[Pos] = n[Which[)];
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(find_first_of(h.begin(), h.end(), n.begin(), n.end()));