--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -181,7 +181,7 @@ _Ty* __std_find_first_of_trivial(
 }
 
 // Can we activate the vector algorithms for find_first_of?
-template <class _It1, class _It2, class _Ty1 = _Iter_value_t<_It1>, class _Ty2 = _Iter_value_t<_It2>>
+template <class _It1, class _It2, class _Pr, class _Ty1 = _Iter_value_t<_It1>, class _Ty2 = _Iter_value_t<_It2>>
 _INLINE_VAR constexpr bool _Vector_alg_in_find_first_of_is_safe =
     _Iterator_is_contiguous<_It1> && _Iterator_is_contiguous<_It2> // contigous
     && !_Iterator_is_volatile<_It1> && !_Iterator_is_volatile<_It2> // non-volatile
@@ -191,7 +191,12 @@ _INLINE_VAR constexpr bool _Vector_alg_in_find_first_of_is_safe =
 #ifdef __cpp_lib_byte
         is_same<_Ty1, byte>, // byte or...
 #endif // defined(__cpp_lib_byte)
-        is_integral<_Ty1>>; // ...integer
+        is_integral<_Ty1>> // ...integer
+    && _Is_any_of_v<_Pr, // predicate is equal
+#if _HAS_CXX20
+        _RANGES equal_to,
+#endif // _HAS_CXX20
+        equal_to<>, equal_to<_Ty1>>;
 _STD_END
 #endif // _USE_STD_VECTOR_ALGORITHMS
 
@@ -3358,8 +3363,7 @@ _NODISCARD _CONSTEXPR20 _FwdIt1 find_first_of(
 
     constexpr ptrdiff_t _Threshold = 16; // vectorization is likely to be a win after this size
 
-    if constexpr (_Vector_alg_in_find_first_of_is_safe<_FwdIt1, _FwdIt2>
-                  && _Is_any_of_v<_Pr, equal_to<>, equal_to<_Ty1>>) {
+    if constexpr (_Vector_alg_in_find_first_of_is_safe<_FwdIt1, _FwdIt2, _Pr>) {
         if (!_STD _Is_constant_evaluated() && _ULast1 - _UFirst1 >= _Threshold) {
             const auto _First1_ptr = _STD _To_address(_UFirst1);
             const auto _Result     = _STD __std_find_first_of_trivial(