--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -180,10 +180,12 @@ _Ty1* __std_find_first_of_trivial(
     }
 }
 
+constexpr ptrdiff_t _Threshold_find_first_of = 16; // find_first_of vectorization is likely to be a win after this size
+
 // Can we activate the vector algorithms for find_first_of?
 template <class _It1, class _It2, class _Pr, class _Ty1 = _Iter_value_t<_It1>, class _Ty2 = _Iter_value_t<_It2>>
 _INLINE_VAR constexpr bool _Vector_alg_in_find_first_of_is_safe =
-    _Iterators_are_contiguous<_It1,_It2> // contiguous
+    _Iterators_are_contiguous<_It1, _It2> // contiguous
     && !_Iterator_is_volatile<_It1> && !_Iterator_is_volatile<_It2> // non-volatile
     && disjunction_v<
 #ifdef __cpp_lib_byte
@@ -3360,10 +3362,8 @@ _NODISCARD _CONSTEXPR20 _FwdIt1 find_first_of(
     const auto _UFirst2 = _STD _Get_unwrapped(_First2);
     const auto _ULast2  = _STD _Get_unwrapped(_Last2);
 #if _USE_STD_VECTOR_ALGORITHMS
-    constexpr ptrdiff_t _Threshold = 16; // vectorization is likely to be a win after this size
-
     if constexpr (_Vector_alg_in_find_first_of_is_safe<_FwdIt1, _FwdIt2, _Pr>) {
-        if (!_STD _Is_constant_evaluated() && _ULast1 - _UFirst1 >= _Threshold) {
+        if (!_STD _Is_constant_evaluated() && _ULast1 - _UFirst1 >= _Threshold_find_first_of) {
             const auto _First1_ptr = _STD _To_address(_UFirst1);
             const auto _Result     = _STD __std_find_first_of_trivial(
                     _First1_ptr, _STD _To_address(_ULast1), _STD _To_address(_UFirst2), _STD _To_address(_ULast2));
@@ -3456,6 +3456,29 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(sentinel_for<_Se2, _It2>);
             _STL_INTERNAL_STATIC_ASSERT(indirectly_comparable<_It1, _It2, _Pr, _Pj1, _Pj2>);
 
+#if _USE_STD_VECTOR_ALGORITHMS
+            if constexpr (_Vector_alg_in_find_first_of_is_safe<_It1, _It2, _Pr> && sized_sentinel_for<_Se1, _It1>
+                          && sized_sentinel_for<_Se2, _It2> && is_same_v<_Pj1, identity> && is_same_v<_Pj2, identity>) {
+                if (!_STD _Is_constant_evaluated() && _Last1 - _First1 >= _Threshold_find_first_of) {
+                    const auto _Count1     = _Last1 - _First1;
+                    const auto _First1_ptr = _STD _To_address(_First1);
+                    const auto _Last1_ptr  = _First1_ptr + _Count1;
+
+                    const auto _Count2     = _Last2 - _First2;
+                    const auto _First2_ptr = _STD _To_address(_First2);
+                    const auto _Last2_ptr  = _First2_ptr + _Count2;
+
+                    const auto _Result =
+                        _STD __std_find_first_of_trivial(_First1_ptr, _Last1_ptr, _First2_ptr, _Last2_ptr);
+
+                    if constexpr (is_pointer_v<_It1>) {
+                        return _Result;
+                    } else {
+                        return _First1 + _Result - _First1_ptr;
+                    }
+                }
+            }
+#endif // _USE_STD_VECTOR_ALGORITHMS
             for (; _First1 != _Last1; ++_First1) {
                 for (auto _Mid2 = _First2; _Mid2 != _Last2; ++_Mid2) {
                     if (_STD invoke(_Pred, _STD invoke(_Proj1, *_First1), _STD invoke(_Proj2, *_Mid2))) {