--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -3350,7 +3350,7 @@ _NODISCARD _CONSTEXPR20 _FwdIt1 find_first_of(
     const auto _UFirst2 = _STD _Get_unwrapped(_First2);
     const auto _ULast2  = _STD _Get_unwrapped(_Last2);
 #if _USE_STD_VECTOR_ALGORITHMS
-    if constexpr (_Vector_alg_in_find_first_of_is_safe<_FwdIt1, _FwdIt2, _Pr>) {
+    if constexpr (_Vector_alg_in_find_first_of_is_safe<decltype(_UFirst1), decltype(_UFirst2), _Pr>) {
         if (!_STD _Is_constant_evaluated() && _ULast1 - _UFirst1 >= _Threshold_find_first_of) {
             const auto _First1_ptr = _STD _To_address(_UFirst1);
             const auto _Result     = _STD __std_find_first_of_trivial(
@@ -3447,7 +3447,7 @@ namespace ranges {
 #if _USE_STD_VECTOR_ALGORITHMS
             if constexpr (_Vector_alg_in_find_first_of_is_safe<_It1, _It2, _Pr> && sized_sentinel_for<_Se1, _It1>
                           && sized_sentinel_for<_Se2, _It2> && is_same_v<_Pj1, identity> && is_same_v<_Pj2, identity>) {
-                if (!_STD _Is_constant_evaluated() && _Last1 - _First1 >= _Threshold_find_first_of) {
+                if (!_STD is_constant_evaluated() && _Last1 - _First1 >= _Threshold_find_first_of) {
                     const auto _Count1     = _Last1 - _First1;
                     const auto _First1_ptr = _STD _To_address(_First1);
                     const auto _Last1_ptr  = _First1_ptr + _Count1;
@@ -3462,7 +3462,7 @@ namespace ranges {
                     if constexpr (is_pointer_v<_It1>) {
                         return _Result;
                     } else {
-                        return _First1 + _Result - _First1_ptr;
+                        return _First1 + (_Result - _First1_ptr);
                     }
                 }
             }