--- a/tests/std/tests/VSO_0000000_vector_algorithms/test.cpp
+++ b/tests/std/tests/VSO_0000000_vector_algorithms/test.cpp
@@ -234,7 +234,7 @@ void test_case_find_first_of(const vector<T>& input_haystack, const vector<T>& i
 template <class T>
 void test_find_first_of(mt19937_64& gen) {
     constexpr size_t needleDataCount = 30;
-    using TD = conditional_t<sizeof(T) == 1, int, T>;
+    using TD                         = conditional_t<sizeof(T) == 1, int, T>;
     uniform_int_distribution<TD> dis('a', 'z');
     vector<T> input_haystack;
     vector<T> input_needle;
@@ -253,11 +253,10 @@ void test_find_first_of(mt19937_64& gen) {
         if (input_haystack.size() == dataCount) {
             break;
         }
-        
+
         input_haystack.push_back(static_cast<T>(dis(gen)));
     }
 }
-
 
 template <class T>
 void test_min_max_element(mt19937_64& gen) {