@AlexGuteniev 2024-03-19
> As we discovered on Discord a while ago, some people call std::find_first_of with a single-element needle. This should be forwarded to std::find (or memchr or something), even if vectorization is otherwise disabled.

I remember this, but it os so much unrelated, that alI think it better fits a separate PR